chore: add scss function 'args-to-list'
Transforms ArgsList to List and appends comma (,) between
different arguments.

Originally, in order to use rfs() mixin on comma seperated values,
the comma sign needs to be included within #{","}.
The args-to-list() function allow to overcome this limitations,
and so it will concatenate all arguments and append comma sign
a comma between them.